update Neo4j driver and fix schema issues
The Neo4j 1.7 driver will be end-of-lifed soon as Neo4j 3.5 becomes EOL
in November 2021. Move to using a newer driver (4.2.1) that still
supports Neo4j 3.5 but also the newer 4.x series.

Fixes and refactors issues with the schema. Schema wasn't being set
before data creation, so a typo in a guid parameter was causing a
collision for nodes representing the "DEFAULT DOMAIN POLICY" and
"DEFAULT DOMAIN CONTROLLERS POLICY" :Base nodes.

The refactor forces the schema to be configured during data creation to
help catch any issues. Clearing the db now just removes the schema
definition but does not recreate it.

This should simplify usage so if users skip "cleardb" on a fresh
database they will still get an index and constraints.